feat: remove-internal-provider-failure-duplicate-fields-003 - Simplify CanonicalWorkFailureMetadata and in-process failure helpers
